begin refactoring Main component to make navigation between servers and channels easier; need to implement auto navigation to 'general' channel for servers
