+++ app/views/api/users/show.json.jbuilder
@@ -64,7 +64,7 @@ json.channels do
     joined_servers.each do |joined_server|
         joined_server.server.channels.each do |channel|
             json.set! channel.id do
-                json.extract! channel, :name, :server_id
+                json.extract! channel, :id, :name, :server_id
             end
         end  
     end
